Add support for Windows x64 (AMD64).
Support has been added for Windows x64, and has been tested on Windows XP x64 only. Builds pass all acceptance tests.

Visual Studio 2008 project files have been added, and the x64 targets have only been added to these files. It should be possible to add the targets to VS2005 project files, but currently only VS2008 is supported.

+++ core/ArrayClass.cpp
@@ -762,7 +762,7 @@ namespace avmplus
 			fieldatoms = NULL;
 		}
 		if(fields) {
-			int numFields = GC::Size(fields)/sizeof(FieldName);
+			int numFields = (int)GC::Size(fields)/sizeof(FieldName);
 			for(int i=0; i<numFields; i++) {
 				fields[i].name = NULL;
 			}
